+++ Changelog.md
@@ -2,6 +2,9 @@
 * Indication added in the BO search line (keywords or space).
 * BO timer deactivated when designing a BO.
 
+## [1.9.2] - 2024.07.09
+* Replace "mouse" with "pynput" implementation (solving Linux issue).
+
 ## [1.9.1] - 2024.03.17
 * Option to keep the BO writer and hotkeys configuration always on top.
 * Add hotkeys to only start OR stop the BO timing run.